backend/requirements.txt
Django==5.0.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
Pillow==10.1.0
djangorestframework-simplejwt==5.3.1
